refactor(Renderer): 패키지 분리(dto -> util)
