+++ src/main/java/chess/controller/ChessController.java
@@ -10,8 +10,8 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.dto.BoardDto;
 import chess.dto.GameRoomDto;
-import chess.dto.PieceRenderer;
-import chess.dto.SquareRenderer;
+import chess.util.PieceRenderer;
+import chess.util.SquareRenderer;
 import chess.view.Command;
 import chess.view.InputView;
 import chess.view.OutputView;
