+++ src/main/java/chess/util/SquareRenderer.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.util;
 
 import chess.domain.board.File;
 import chess.domain.board.Rank;
